Added metadata to imported assets
The "Import into Content Browser" loader app action will add metadata for "created_by" and the PublishedFile "url" to the imported asset in Unreal. The imported asset will get focus in the Content Browser after a successful import.

The "Jump to Shotgun" action in the context menus will now jump to the URL specified in the metadata of the selected actor/asset if present. Otherwise, it will jump to the project URL.