Added command override to launch Shotgun Panel with selected context
An asset previously imported with the Loader app will contain the URL metadata that can provide a context usable with Shotgun apps. In this case, the Shotgun Panel command has been overridden to launch the Shotgun Panel with the context of the asset selected in the Unreal Content Browser or the actor selected in the Level Editor/World Outliner.